Developing BinaryGalprop into a full-scale feature
